refactor: rename MapSidebar to Sidebar
